Splitted the ProductPage file
